ax_enums.mojom 1
Explicitly casting all uses of enums from ax_enums.idl. This is
preparatory work so we can convert these enums to mojo in a follow-up.

Bug: 650275

+++ ash/system/night_light/night_light_toggle_button.cc
@@ -80,7 +80,8 @@ void NightLightToggleButton::GetAccessibleNodeData(ui::AXNodeData* node_data) {
   node_data->role = ui::AX_ROLE_TOGGLE_BUTTON;
   node_data->AddIntAttribute(
       ui::AX_ATTR_CHECKED_STATE,
-      is_enabled ? ui::AX_CHECKED_STATE_TRUE : ui::AX_CHECKED_STATE_FALSE);
+      static_cast<int32_t>(is_enabled ? ui::AX_CHECKED_STATE_TRUE
+                                      : ui::AX_CHECKED_STATE_FALSE));
 }
 
 }  // namespace ash
